refactor(simAction): move chat parser into chatMessageActions
