+++ src/actions/simAction.js
@@ -1,64 +1,20 @@
-import {createNewIMChat} from './chatMessageActions'
-import {
-  getValueOf,
-  getStringValueOf,
-  mapBlockOf
-} from '../network/msgGetters'
+import {createNewIMChat, parseChatFromSimulator, parseIM} from './chatMessageActions'
+import {getValueOf, mapBlockOf} from '../network/msgGetters'
 
-function parseChatFromSimulator (msg) {
-  const chatMsg = {
-    fromName: getStringValueOf(msg, 'ChatData', 'FromName'),
-    sourceID: getValueOf(msg, 'ChatData', 'SourceID'),
-    ownerID: getValueOf(msg, 'ChatData', 'OwnerID'),
-    sourceType: getValueOf(msg, 'ChatData', 'SourceType'),
-    chatType: getValueOf(msg, 'ChatData', 'ChatType'),
-    audible: getValueOf(msg, 'ChatData', 'Audible'),
-    position: getValueOf(msg, 'ChatData', 'Position'),
-    message: getStringValueOf(msg, 'ChatData', 'Message'),
-    time: Date.now()
-  }
-  return chatMsg
-}
-
-function parseIM (message) {
-  const toAgentID = getValueOf(message, 'MessageBlock', 'ToAgentID')
-  const fromId = getValueOf(message, 'AgentData', 'AgentID')
-  const time = getValueOf(message, 'MessageBlock', 'Timestamp')
-
-  const IMmsg = {
-    sessionID: getValueOf(message, 'AgentData', 'SessionID'),
-    fromId,
-    fromGroup: getValueOf(message, 'MessageBlock', 'FromGroup'),
-    toAgentID,
-    parentEstateID: getValueOf(message, 'MessageBlock', 'ParentEstateID'),
-    regionID: getValueOf(message, 'MessageBlock', 'RegionID'),
-    position: getValueOf(message, 'MessageBlock', 'Position'),
-    offline: getValueOf(message, 'MessageBlock', 'Offline'),
-    dialog: getValueOf(message, 'MessageBlock', 'Dialog'),
-    id: getValueOf(message, 'MessageBlock', 'ID'),
-    fromAgentName: getStringValueOf(message, 'MessageBlock', 'FromAgentName'),
-    message: getStringValueOf(message, 'MessageBlock', 'Message'),
-    binaryBucket: getValueOf(message, 'MessageBlock', 'BinaryBucket'),
-    time: time !== 0 ? time : Date.now()
-  }
-
-  // If it is a group chat, toAgentID is the Group-UUID.
-  IMmsg.chatUUID = IMmsg.fromGroup ? IMmsg.toAgentID : IMmsg.id
-  return IMmsg
-}
-
-function parseUserRights (message, getState) {
-  const rights = mapBlockOf(message, 'Rights', getValue => {
-    return {
-      agentId: getValue('AgentRelated'),
-      rights: getValue('RelatedRights')
-    }
-  })
-  return {
-    type: 'ChangeUserRights',
-    ownId: getState().account.get('agentId'),
-    fromId: getValueOf(message, 'AgentData', 'AgentID'),
-    userRights: rights
+function parseUserRights (message) {
+  return (dispatch, getState) => {
+    const rights = mapBlockOf(message, 'Rights', getValue => {
+      return {
+        agentId: getValue('AgentRelated'),
+        rights: getValue('RelatedRights')
+      }
+    })
+    dispatch({
+      type: 'ChangeUserRights',
+      ownId: getState().account.get('agentId'),
+      fromId: getValueOf(message, 'AgentData', 'AgentID'),
+      userRights: rights
+    })
   }
 }
 
@@ -69,7 +25,7 @@ function simActionFilter (msg) {
   switch (name) {
     case 'ChatFromSimulator':
       const parsed = parseChatFromSimulator(msg)
-      return dispatchSIMAction(name, parsed, 'localchat/' + new Date(parsed.time).toJSON())
+      return dispatchChatAction(name, parsed, 'localchat/' + new Date(parsed.time).toJSON())
 
     case 'ImprovedInstantMessage':
       const parsedMsg = parseIM(msg)
@@ -79,13 +35,11 @@ function simActionFilter (msg) {
           parsedMsg.dialog, parsedMsg.chatUUID, parsedMsg.fromId, parsedMsg.fromAgentName
         ))
         const id = `imChats/${parsedMsg.chatUUID}/${new Date(parsedMsg.time).toJSON()}`
-        dispatch(dispatchSIMAction(name, parsedMsg, id))
+        dispatch(dispatchChatAction(name, parsedMsg, id))
       }
 
     case 'ChangeUserRights':
-      return (dispatch, getState) => {
-        dispatch(parseUserRights(msg, getState))
-      }
+      return parseUserRights(msg)
 
     case 'RegionHandshake':
       return sendRegionHandshakeReply(msg)
@@ -101,20 +55,23 @@ function simActionFilter (msg) {
   }
 }
 
-// Dispatches all parsed messages.
-// If they have an ID, they will be saved and synced under the avatar name.
-function dispatchSIMAction (name, msg, id) {
-  return (dispatch, getState, {hoodie}) => {
+// Dispatches chat (and IM) messages.
+// They will be saved and synced under the avatar name.
+function dispatchChatAction (name, msg, id) {
+  return async (dispatch, getState, {hoodie}) => {
+    if (typeof id !== 'string') return
+
     const activeState = getState()
-    if (typeof id === 'string' && activeState.account.getIn(['viewerAccount', 'loggedIn'])) {
+    if (
+      activeState.account.getIn(['viewerAccount', 'loggedIn']) && activeState.account.get('sync')
+    ) {
       // Save messages. They will also be synced!
-      const avatarIdentifier = activeState.account.get('avatarIdentifier')
-      msg._id = avatarIdentifier + '/' + id
-      hoodie.store.add(msg).then(doc => {
-        dispatch({
-          type: name,
-          msg: doc
-        })
+      msg._id = activeState.account.get('avatarIdentifier') + '/' + id
+
+      const doc = await hoodie.store.add(msg)
+      dispatch({
+        type: name,
+        msg: doc
       })
     } else {
       // This is the path for every message, that will not be synced and saved.
